refactor: model and repository
